backend/requirements.txt
numpy==1.13.3
opencv_python==4.1.1.26
Pillow==7.0.0
pymongo==3.10.1
Flask==1.1.1
flask_cors==3.0.8
flask_jwt==0.3.2
glob2==0.7
lxml==4.5.0
pyftpdlib==1.5.6
pyunpack==0.1.2
schedule==0.6.0
Werkzeug==1.0.0
